backend/requirements.txt
# Core framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Auth & Security
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.5.0
cryptography==46.0.5

# Pydantic & validation
pydantic==2.12.5
pydantic_core==2.41.5
email-validator==2.3.0

# Env & config
python-dotenv==1.2.1

# File upload support
python-multipart==0.0.22

# Payment
razorpay==2.0.0

# HTTP utilities (fastapi/httpx transitive deps)
anyio==4.12.1
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1
idna==3.11
certifi==2026.1.4
requests==2.32.5
urllib3==2.6.3

# Pydantic runtime
annotated-types==0.7.0
typing_extensions==4.15.0
